# Conflicts:
#	app/.gitignore
#	app/build.gradle.kts
#	app/secrets.properties
#	app/src/main/assets/databases/initial_db.db
#	app/src/main/java/com/f4/mypet/AppModule.kt
#	app/src/main/java/com/f4/mypet/MainActivity.kt
#	app/src/main/java/com/f4/mypet/data/db/DBRepository.kt
#	app/src/main/java/com/f4/mypet/data/network/ApiHandler.kt
#	app/src/main/java/com/f4/mypet/data/network/NetworkRepository.kt
#	app/src/main/java/com/f4/mypet/data/network/NetworkRepositoryImpl.kt
#	app/src/main/java/com/f4/mypet/data/network/model/NetworkResult.kt
#	app/src/main/java/com/f4/mypet/data/network/service/MedRecordService.kt
#	app/src/main/java/com/f4/mypet/data/network/service/ProcedureService.kt
#	app/src/main/java/com/f4/mypet/ui/screens/ErrorScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/LoadingScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/auth/login/LoginScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/medcard/createUpdate/success/screenComponents/SaveButton.kt
#	app/src/main/java/com/f4/mypet/ui/screens/medcard/list/ListMedRecords.kt
#	app/src/main/java/com/f4/mypet/ui/screens/medcard/list/success/MedRecordItem.kt
#	app/src/main/java/com/f4/mypet/ui/screens/medcard/show/MedRecordScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/medcard/show/screenComponents/RemoveMedRecordAlert.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/createUpdate/CreateUpdateProcedureViewModel.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/createUpdate/success/SuccessCUProcedureScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/list/ListProcedureScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/list/ListProcedureViewModel.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/list/success/ProcedureItem.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/list/success/SuccessListProcedureScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/show/ProcedureScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/show/ProcedureViewModel.kt
#	app/src/main/java/com/f4/mypet/ui/screens/procedure/show/RemoveProcedureAlert.kt
#	app/src/main/java/com/f4/mypet/ui/screens/profile/createUpdate/CreateUpdateProfileScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/profile/list/ListProfileViewModel.kt
#	app/src/main/java/com/f4/mypet/ui/screens/profile/list/success/SuccessListProfileScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/profile/show/ProfileScreen.kt
#	app/src/main/java/com/f4/mypet/ui/screens/profile/show/ProfileViewModel.kt
#	app/src/main/java/com/f4/mypet/util/Validations.kt

+++ app/src/main/java/com/f4/mypet/ui/screens/profile/show/ProfileViewModel.kt
@@ -1,10 +1,14 @@
 package com.f4.mypet.ui.screens.profile.show
 
+import android.content.Context
+import android.content.Intent
+import android.widget.Toast
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.f4.mypet.data.db.Repository
 import com.f4.mypet.data.db.entities.Pet
 import com.f4.mypet.data.network.NetworkRepository
+import com.f4.mypet.data.network.model.NetworkResult
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,6 +17,7 @@ import kotlinx.coroutines.launch
 import java.time.LocalDate
 import javax.inject.Inject
 
+@SuppressWarnings("TooGenericExceptionCaught")
 @HiltViewModel
 class ProfileViewModel @Inject constructor(
     private val repository: Repository,
@@ -34,9 +39,8 @@ class ProfileViewModel @Inject constructor(
     fun getPetProfile(petId: Int) {
         viewModelScope.launch(Dispatchers.IO) {
             if (petId != -1) {
+                // получение локально, т.к. данные обновились при получении списка питомцев
                 _petUiState.value = repository.getPet(petId)
-                // пока подтягивание с сервера отключено
-//                _petUiState.value = networkRepository.getPet(petId) ?: _petUiState.value
             }
         }
     }
@@ -44,10 +48,36 @@ class ProfileViewModel @Inject constructor(
 
     fun removePet(pet: Pet) {
         viewModelScope.launch(Dispatchers.IO) {
-            repository.removePet(pet)
-            repository.removeProceduresForPet(pet.id)
-            repository.removeMedRecordsForPet(pet.id)
-            _msg.value = networkRepository.removePet(pet.id)
+            when (val response = networkRepository.removePet(pet.id)) {
+                is NetworkResult.Success -> {
+                    _msg.value = null
+                    repository.removePet(pet)
+                    repository.removeProceduresForPet(pet.id)
+                    repository.removeMedRecordsForPet(pet.id)
+                }
+                is NetworkResult.Error -> { _msg.value = response.msg }
+            }
         }
+    }
+
+    fun sharePetInfo(message: String, context: Context) {
+        val intent = Intent(Intent.ACTION_SEND).apply {
+            putExtra(Intent.EXTRA_TEXT, message)
+            type = "text/plain"
+        }
+        try {
+            context.startActivity(
+                Intent.createChooser(
+                    intent,
+                    "Отправить сведения о питомце"
+                )
+            )
+        } catch (e: Exception) {
+            Toast.makeText(context, "Произошла ошибка", Toast.LENGTH_LONG).show()
+        }
+    }
+
+    fun resetMsg() {
+        _msg.value = ""
     }
 }
